New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Fix of viewer crash when multiple images were loaded, close All clicked, and profiler open.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@24133 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/QtPlotter/QtProfile.qo.h
@@ -200,7 +200,7 @@ public slots:
 	void changeSpectrum(String spcTypeUnit, String spcRval, String spcSys);
 
 	void plotMainCurve();
-	void setCollapseRange(float xmin, float xmax);
+	void setCollapseRange(double xmin, double xmax);
 
 	void overplot(QHash<QString, ImageInterface<float>*>);
